Crate ixgbe

An ixgbe driver for a 82599 10GbE Network Interface Card.

Currently we support basic send and receive, Receive Side Scaling (RSS), 5-tuple filters, and MSI interrupts. We also support language-level virtualization of the NIC so that applications can directly access their assigned transmit and receive queues. When using virtualization, we disable RSS since we use 5-tuple filters to ensure packets are routed to the correct queues. We also disable interrupts when using virtualization, since we do not yet have support for allowing applications to register their own interrupt handlers.

Modules

test_packets
A set of functions to create packets for testing the NIC transmission functionality.

virtual_function
Interface for an application to request a VirtualNIC from the ixgbe device, and implementation of the PhysicalNic trait for the ixgbe device.

Structs

IxgbeNic
A struct representing an ixgbe network interface card.

Enums

FilterProtocol
Options for the filter protocol used in the 5-tuple filters.

LinkSpeedMbps
Possible link speeds of the 82599 NIC

RxBufferSizeKiB
The set of receive buffer sizes that are accepted by the 82599 device.

Constants

INTEL_82599
Device ID for the 82599ES, used to identify the device from the PCI space

INTEL_VEND
Vendor ID for Intel

Statics

IXGBE_NICS
All the 82599 NICs found in the PCI space are initialized and then stored here.

Functions

get_ixgbe_nic
Returns a reference to the IxgbeNic wrapped in a MutexIrqSafe, if it exists and has been initialized. Currently we use the pci location of the device as identification since it should not change after initialization.

get_ixgbe_nics_list
Returns a reference to the list of all initialized ixgbe NICs